+++ examples/zent-modulith/src/modules/catalog/persistence.zig
@@ -14,6 +14,7 @@ const graph = zent.codegen.graph.buildGraph(&.{
     model.Inventory,
     model.Order,
     model.Account,
+    model.SkuStock,
 });
 pub const infos = graph.types;
 pub const Client = zent.codegen.client.Client(infos);
@@ -26,6 +27,7 @@ pub const CommentInfo = infos[6];
 pub const InventoryInfo = infos[7];
 pub const OrderInfo = infos[8];
 pub const AccountInfo = infos[9];
+pub const SkuStockInfo = infos[10];
 
 pub const CatalogStore = struct {
     allocator: std.mem.Allocator,